+++ src/JsonMapperInterface.php
@@ -7,4 +7,6 @@ namespace DannyVanDerSluijs\JsonMapper;
 interface JsonMapperInterface
 {
     public function mapObject(\stdClass $json, object $object): void;
+
+    public function mapArray(\stdClass $json, object $object): array;
 }